Shifted filter tokenizer into a new class
